refactor(common,starknet_api): move test util and transaction_hash_test into starknet_api
