+++ crates/papyrus_rpc/src/compression_utils_test.rs
@@ -1,5 +1,5 @@
-use papyrus_test_utils::read_json_file;
 use pretty_assertions::assert_eq;
+use starknet_api::test_utils::read_json_file;
 
 use super::compress_and_encode;
