+++ crates/starknet_api/Cargo.toml
@@ -17,6 +17,7 @@ hex.workspace = true
 indexmap = { workspace = true, features = ["serde"] }
 itertools.workspace = true
 num-bigint.workspace = true
+pretty_assertions.workspace = true
 primitive-types = { workspace = true, features = ["serde"] }
 serde = { workspace = true, features = ["derive", "rc"] }
 serde_json.workspace = true
