+++ crates/starknet_api/src/test_utils.rs
@@ -1,4 +1,7 @@
 use std::collections::HashMap;
+use std::env;
+use std::fs::read_to_string;
+use std::path::Path;
 
 use starknet_types_core::felt::Felt;
 
@@ -11,6 +14,16 @@ pub mod invoke;
 #[derive(Default)]
 pub struct NonceManager {
     next_nonce: HashMap<ContractAddress, Felt>,
+}
+
+/// Reads from the directory containing the manifest at run time, same as current working directory.
+pub fn read_json_file(path_in_resource_dir: &str) -> serde_json::Value {
+    let path = Path::new(&env::var("CARGO_MANIFEST_DIR").unwrap())
+        .join("resources")
+        .join(path_in_resource_dir);
+    let json_str = read_to_string(path.to_str().unwrap())
+        .expect(format!("Failed to read file at path: {}", path.display()).as_str());
+    serde_json::from_str(&json_str).unwrap()
 }
 
 impl NonceManager {
