+++ crates/starknet_api/src/transaction_hash_test.rs
@@ -1,10 +1,6 @@
-use papyrus_test_utils::read_json_file;
 use pretty_assertions::assert_eq;
 use serde::{Deserialize, Serialize};
 use sha3::{Digest, Keccak256};
-use starknet_api::block::BlockNumber;
-use starknet_api::core::ChainId;
-use starknet_api::transaction::{Transaction, TransactionHash, TransactionOptions};
 use starknet_types_core::felt::Felt;
 
 use super::{
@@ -13,6 +9,10 @@ use super::{
     validate_transaction_hash,
     CONSTRUCTOR_ENTRY_POINT_SELECTOR,
 };
+use crate::block::BlockNumber;
+use crate::core::ChainId;
+use crate::test_utils::read_json_file;
+use crate::transaction::{Transaction, TransactionHash, TransactionOptions};
 
 #[test]
 fn test_ascii_as_felt() {
@@ -30,7 +30,7 @@ fn test_constructor_selector() {
     let mut constructor_bytes: [u8; 32] = keccak.finalize().into();
     constructor_bytes[0] &= 0b00000011_u8; // Discard the six MSBs.
     let constructor_felt = Felt::from_bytes_be(&constructor_bytes);
-    assert_eq!(constructor_felt, *CONSTRUCTOR_ENTRY_POINT_SELECTOR);
+    assert_eq!(constructor_felt, CONSTRUCTOR_ENTRY_POINT_SELECTOR);
 }
 
 #[derive(Deserialize, Serialize)]
